+++ internal/wallet/wallet/lib.go
@@ -1,4 +1,4 @@
-package lib
+package wallet
 
 import (
 	"fmt"
