+++ internal/wallet/wallet/lib_test.go
@@ -1,4 +1,4 @@
-package lib
+package wallet
 
 import (
 	"testing"
